fix: merge `try_lockup` methods
